[BZ202025] Delivering fix to disconnect before deleting profiles

+++ plugins/org.eclipse.datatools.connectivity/src/org/eclipse/datatools/connectivity/internal/InternalProfileManager.java
@@ -33,7 +33,9 @@ import org.eclipse.core.resources.IResource;
 import org.eclipse.core.resources.ResourcesPlugin;
 import org.eclipse.core.runtime.CoreException;
 import org.eclipse.core.runtime.FileLocator;
+import org.eclipse.core.runtime.IStatus;
 import org.eclipse.core.runtime.ListenerList;
+import org.eclipse.core.runtime.Status;
 import org.eclipse.datatools.connectivity.ConnectionProfileException;
 import org.eclipse.datatools.connectivity.ICategory;
 import org.eclipse.datatools.connectivity.IConnectionProfile;
@@ -548,7 +550,14 @@ public class InternalProfileManager {
 	 */
 	public void deleteProfile(IConnectionProfile profile)
 			throws ConnectionProfileException {
-		IConnectionProfileRepository repo = getRepositoryForProfile(profile);
+	    
+	    IStatus status = profile.disconnect();
+	    if (status == Status.CANCEL_STATUS)
+        {
+            return;
+        }
+	    
+	    IConnectionProfileRepository repo = getRepositoryForProfile(profile);
 		if (repo != null) {
 			repo.deleteProfile(profile);
 			return;
@@ -577,8 +586,7 @@ public class InternalProfileManager {
 				.toArray(new IConnectionProfile[0]);
 
 		mIsDirty = true;
-
-		profile.disconnect();
+		
 		fireProfileDeleted(profile);
 		saveChanges();
 	}
